Removed Internal Array Table
I took this out because I was using it as a crutch in lieu of not knowing how to traverse the tree properly. If one knows how to traverse properly there is no need for a internal backing array, because the Tree is really an extension of LinkedList. What makes it special is the O(log n) search that can be achieved.